WIP for User Stories US07192025309PM and US07192025306PM
Cleaned project in accordance to MVC architecture.
